moved class to separate file and renamed
